backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
tzdata   # si necesitás datos de zona horaria
